refactored current implementation into new project structure.
